Refactor WalletConnect integration: move WalletConnect logic to a separate module and enhance type definitions
